refactor: rename packageimport.Registry to RequestManager
having a struct definition for `Registry` in a file called registry.go, which is called like that because it contains a `FromRegistry` function that pulls a package image from an OCI registry (just like oci.go contains a `FromOCI` function!) although that `Registry` struct is actually a facility to deduplicate concurrent pull requests for the same package image.... yeah that confuses my little smoothbrain way too much
